+++ BeetMall/src/main/java/com/beetmall/sshj/custom/vo/ReviewVO.java
@@ -2,7 +2,7 @@ package com.beetmall.sshj.custom.vo;
 
 public class ReviewVO {
 	private int reviewnum;
-	private long ordernum;
+	private int ordernum;
 	private int productnum;
 	private String reviewcontent;
 	private String userid;
@@ -26,10 +26,10 @@ public class ReviewVO {
 	public void setReviewnum(int reviewnum) {
 		this.reviewnum = reviewnum;
 	}
-	public long getOrdernum() {
+	public int getOrdernum() {
 		return ordernum;
 	}
-	public void setOrdernum(long ordernum) {
+	public void setOrdernum(int ordernum) {
 		this.ordernum = ordernum;
 	}
 	public int getProductnum() {
